+++ workflow/scripts/csaw/csaw_count.R
@@ -66,3 +66,9 @@ base::saveRDS(
     object = snakemake@output[["rds"]]
 )
 base::message("Process over")
+
+
+# Proper syntax to close the connection for the log file
+# but could be optional for Snakemake wrapper
+base::sink(type = "message");
+base::sink();