+++ workflow/scripts/csaw/csaw_edger.R
@@ -75,3 +75,8 @@ if ("edger" %in% base::names(snakemake@ouptput)) {
     base::message("edgeR results saved")
 }
 base::message("Process over")
+
+# Proper syntax to close the connection for the log file
+# but could be optional for Snakemake wrapper
+base::sink(type = "message");
+base::sink();