+++ workflow/scripts/csaw/csaw_fdr.R
@@ -101,3 +101,8 @@ if ("csv" %in% base::names(snakemake@output)) {
     )
 }
 base::message("Process over")
+
+# Proper syntax to close the connection for the log file
+# but could be optional for Snakemake wrapper
+base::sink(type = "message");
+base::sink();