+++ workflow/scripts/csaw/csaw_filter.R
@@ -141,3 +141,8 @@ base::saveRDS(
     object = counts,
     file = base::as.character(x = snakemake@output[["rds"]])
 )
+
+# Proper syntax to close the connection for the log file
+# but could be optional for Snakemake wrapper
+base::sink(type = "message");
+base::sink();