+++ workflow/scripts/csaw/csaw_readparam.R
@@ -38,3 +38,8 @@ base::saveRDS(
     object = read_params,
     file = snakemake@output[["rds"]]
 )
+
+# Proper syntax to close the connection for the log file
+# but could be optional for Snakemake wrapper
+base::sink(type = "message");
+base::sink();